refactor kubernetes redirect code
